Moved ability routes
