+++ src/server/src/ability/controller.ts
@@ -1,0 +1,12 @@
+import { getAllAbilityMetadata as getAllAbilityMetadataService } from 'src/services/getAllAbilityMetadata';
+import { asyncHandler } from 'src/asyncHandler';
+
+const getAllAbilityMetadata = asyncHandler(async (req, res, next) => {
+  const results = await getAllAbilityMetadataService();
+
+  res.json(results);
+});
+
+export const controller = {
+  getAllAbilityMetadata,
+};
