+++ src/server/src/ability/router.ts
@@ -1,0 +1,6 @@
+import { Router } from 'express';
+import { controller } from 'src/ability/controller';
+
+export const router = Router();
+
+router.get('/ability', controller.getAllAbilityMetadata);
