+++ src/server/src/router.ts
@@ -2,6 +2,8 @@ import { Router } from 'express';
 import { controller } from 'src/controller';
 import { router as playerRouter } from 'src/player/router';
 import { router as leaderboardRouter } from 'src/leaderboard/router';
+import { router as itemRouter } from 'src/item/router';
+import { router as abilityRouter } from 'src/ability/router';
 
 export const router = Router();
 
@@ -11,10 +13,8 @@ router.get('/status', (req, res, next) => {
 
 router.use(playerRouter);
 router.use(leaderboardRouter);
-
-router.get('/item', controller.getItemMetadata);
-
-router.get('/ability', controller.getAllAbilityMetadata);
+router.use(itemRouter);
+router.use(abilityRouter);
 
 router.get('/search/player', controller.searchPlayer);
